util/python/requirements.txt
pika==0.12
mysql-connector